finished the functionality

+++ window-shopper/Model/Wage.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 class Wage {
-    class func getHours(forWage wage: Double, andPrice price: Double) -> Double {
-        return ceil(price / wage)
+    class func getHours(forWage wage: Double, andPrice price: Double) -> Int {
+        return Int(ceil(price / wage))
     }
 }
